Move converters to module
