+++ app/build.gradle
@@ -28,4 +28,5 @@ dependencies {
 
     api "com.jakewharton:butterknife:8.8.1"
     annotationProcessor "com.jakewharton:butterknife-compiler:8.8.1"
+    implementation project(":converters")
 }
