+++ converters/src/main/java/com/example/converters/Converters.java
@@ -1,4 +1,4 @@
-package com.example.evgeniy.htcandroid;
+package com.example.converters;
 
 import android.content.Context;
